+++ DICOM/Imaging/LUT/ModalityLUT.cs
@@ -6,9 +6,7 @@ namespace Dicom.Imaging.LUT {
 	/// </summary>
 	public class ModalityLUT : ILUT {
 		#region Private Members
-		private double _rescaleSlope;
-		private double _rescaleIntercept;
-
+		private GrayscaleRenderOptions _renderOptions;
 		private int _minValue;
 		private int _maxValue;
 		#endregion
@@ -17,15 +15,11 @@ namespace Dicom.Imaging.LUT {
 		/// <summary>
 		/// Initialize new instance of <seealso cref="ModalityLUT"/> using the specified slope and intercept parameters
 		/// </summary>
-		/// <param name="minValue">The minimum input pixel value</param>
-		/// <param name="maxValue">The maximum input pixel value</param>
-		/// <param name="slope">The modality LUT rescale slope</param>
-		/// <param name="intercept">The modality LUT rescale intercept</param>
-		public ModalityLUT(int minValue, int maxValue, double slope, double intercept) {
-			_rescaleSlope = slope;
-			_rescaleIntercept = intercept;
-			_minValue = this[minValue];
-			_maxValue = this[maxValue];
+		/// <param name="options">Render options</param>
+		public ModalityLUT(GrayscaleRenderOptions options) {
+			_renderOptions = options;
+			_minValue = this[options.BitDepth.MinimumValue];
+			_maxValue = this[options.BitDepth.MaximumValue];
 		}
 		#endregion
 
@@ -34,14 +28,14 @@ namespace Dicom.Imaging.LUT {
 		/// The modality rescale slope
 		/// </summary>
 		public double RescaleSlope {
-			get { return _rescaleSlope; }
+			get { return _renderOptions.RescaleSlope; }
 		}
 
 		/// <summary>
 		/// The modality rescale intercept
 		/// </summary>
 		public double RescaleIntercept {
-			get { return _rescaleIntercept; }
+			get { return _renderOptions.RescaleIntercept; }
 		}
 
 		public bool IsValid {
@@ -58,7 +52,7 @@ namespace Dicom.Imaging.LUT {
 
 		public int this[int value] {
 			get {
-				return unchecked((int)((value * _rescaleSlope) + _rescaleIntercept));
+				return unchecked((int)((value * RescaleSlope) + RescaleIntercept));
 			}
 		}
 		#endregion