+++ ChangeLog.md
@@ -8,6 +8,7 @@
 * Fix Async Ops window for associations where it is not negotiated
 * Fix bug reading Palette Color LUT with implicit length
 * Support decompression and rendering of JPEG Process 2 images
+* Enable modification of Window/Level in DicomImage
 
 #### v1.0.26 (10/19/2012)
 * Advanced DIMSE logging options for DicomService base class